Core data settings

+++ fitnessmebel/Accessory+CoreDataProperties.h
@@ -16,9 +16,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSFetchRequest<Accessory *> *)fetchRequest;
 
-@property (nonatomic) int32_t innerID;
-@property (nonatomic) int32_t remoteID;
-@property (nonatomic) int32_t sequence;
+@property (nonatomic) int64_t innerID;
+@property (nonatomic) int64_t remoteID;
+@property (nonatomic) int64_t sequence;
 @property (nullable, nonatomic, copy) NSString *text;
 @property (nullable, nonatomic, copy) NSString *thumbnailFolderPath;
 @property (nullable, nonatomic, copy) NSString *thumbnailURLPath;
