+++ fitnessmebel/Gallery+CoreDataProperties.h
@@ -16,10 +16,10 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSFetchRequest<Gallery *> *)fetchRequest;
 
-@property (nonatomic) int32_t furnitureID;
-@property (nonatomic) int32_t innerID;
-@property (nonatomic) int32_t remoteID;
-@property (nonatomic) int32_t sequence;
+@property (nonatomic) int64_t furnitureID;
+@property (nonatomic) int64_t innerID;
+@property (nonatomic) int64_t remoteID;
+@property (nonatomic) int64_t sequence;
 @property (nullable, nonatomic, copy) NSString *thumbnailFolderPath;
 @property (nullable, nonatomic, copy) NSString *thumbnailURLPath;
 @property (nullable, nonatomic, copy) NSString *title;
